Update restrict-your-spot-iam-policy.md
A customer reached out with a problem with the restricted VPC. 
He tried to user restrict VPC to a Cloudwatch resource but he received an error. After investigation, we found out that restrict VPC is not relevant for all of the AWS resources.
So I want to clarify in the documentation that the example is for EC2 resources.